+++ Solfeggio/Solfeggio/Api/IWave.cs
@@ -18,7 +18,7 @@ namespace Solfeggio.Api
 
 	public interface IProcessor : IDataSource
 	{
-		float Level { get; set; }
+		double Level { get; set; }
 		double Boost { get; set; }
 		void Wake();
 		void Lull();
